requirements.txt
alembic==1.7.1
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.1
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-resources==5.2.2
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
zipp==3.5.0
